refactor: move openapi client generation into separate directory
