--- docs/client/package.json
@@ -1,21 +1,0 @@
-{
-  "name": "@blockstack/client",
-  "version": "1.0.0",
-  "description": "",
-  "main": "index.js",
-  "scripts": {
-    "build": "rimraf ./lib && tsc",
-    "test": "ts-node src/test.ts"
-  },
-  "keywords": [],
-  "author": "",
-  "license": "ISC",
-  "devDependencies": {
-    "rimraf": "^3.0.2",
-    "ts-node": "^8.10.2",
-    "typescript": "^3.9.7"
-  },
-  "dependencies": {
-    "cross-fetch": "^3.0.5"
-  }
-}
