--- docs/client/tsconfig.json
@@ -1,10 +1,0 @@
-{
-  "compilerOptions": {
-    "target": "ES2015",
-    "module": "commonjs",
-    "outDir": "./lib",
-    "strict": true,
-    "baseUrl": "./"
-  },
-  "include": ["src/**/*"],
-}
